Enable geo_distance and geo_bounding_box queries on geo_shape field type (#64224)
